app-redator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
anthropic==0.42.0
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic[email]==2.10.3
python-multipart==0.0.20
requests>=2.31.0
psycopg2-binary>=2.9.9
